Adds a delivery mode called async-deferred
I added a new result delivery mode called async-deferred, this mode returns a handle once the query completed.
context: sync is a blocking call which returns the result once the query completed, whereas async returns a handle immediately to pull the results later.
async-defer is a hybrid of these two as it returns a handle but only once the query completed.

Change-Id: I8480d22ea9b107edf6d340073581c2613b017599
Reviewed-on: http://fulliautomatix.ics.uci.edu:8443/77
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Till Westmann <westmann@gmail.com>