Fix issue 897 (Stale metadata breaks ADM/AQL 101)
This fixes the issue for me, but I think this really needs some careful consideration.
Identical issues could be lurking anywhere the passed object is modified and that info needs to be conveyed to the caller.
I feel like I should not have to read from MetadataNode directly to get this info.

Change-Id: I088d33ad7810d7e1ca5471202cf20fa9a651195c
Reviewed-on: https://asterix-gerrit.ics.uci.edu/283
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ildar Absalyamov <ildar.absalyamov@gmail.com>