Change folder structure for Java repackage
Change only the folders, not the files, for our package name change.
This will break the build, and needs to be followed by a change to
the package name in all of the source files. However performing
the folder move and file change in two steps lets Git understand
that the files are the same, and lets us track revisions across
those files.

Change-Id: Iefd2a576415ebc1416cba2a3334d2b64f042ba92
Reviewed-on: https://asterix-gerrit.ics.uci.edu/306
Tested-by: Ian Maxon <imaxon@apache.org>
Reviewed-by: Till Westmann <tillw@apache.org>