ASTERIXDB-1148: Selectable array-wrapping of results
Introduce "wrapper-array" parameter to HTTP API which selects (for ADM and
JSON) whether to wrap the result sequence in a generated outer array. For
JSON this defaults to "true" as before. For ADM this defaults to false,
resulting in a large number of expected-results changes.

Also introduce ability to have AQL tests which provide HTTP parameters.

Change-Id: I3122f136ff9ca8a2c2268238c57bb5eddab7b27e
Reviewed-on: https://asterix-gerrit.ics.uci.edu/473
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Chris Hillery <ceej@lambda.nu>