Run ExecutionTest w/integration test framework
I changed the default local.xml to use 2 NCs with
2 partitions each like AsterixHyracksIntegrationUtil
so that the results will match despite using simple
string compare to determine correctness.

There is also one test in particular (big-object-join)
that I had to add an artificial order by clause to
to deal with different results.

Issues outstanding:
- HDFS in ClusterExecutionIT needs to be thought out

Change-Id: I423f2a7c77839b999d466dd5cace302574d956c0
Reviewed-on: https://asterix-gerrit.ics.uci.edu/525
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Chris Hillery <ceej@lambda.nu>