ASTERIXDB-1302 ASTERIXDB-1301 Fix Socket Feed Connection
A bug causes a read lock to never be released when a feed is
connected with "wait-for-completion" set to false. The bug
was fixed and a test case was added.
Another bug was causing the socket feed to not receive
connections correctly. The bug was fixed and a test case
was added.
Additionally, this change ensures that adapters have absolute
partitions to ensure consistency with regards to feed log
manager.

Change-Id: I8f6e982440d3577343f2479c3779653a9c3db614
Reviewed-on: https://asterix-gerrit.ics.uci.edu/660
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ildar Absalyamov <ildar.absalyamov@gmail.com>