Fixes for ASTERIXDB-1636
The index of the tuple field for filters from SecondaryIndexOperationsHelper and AqlMetadataProvider
differed. The one in AqlMetadataProvider was wrong, as it was attempting to take into account the
presence of a partitioning field in the incoming tuple, which is not there in the case of an
insert/upsert.

There was also an issue where on merge, for components with a filter page but no min/max, the merge
would fail. I fixed this by skipping over null entries while getting the min/max of merging components.

Finally, there was a very silly error in LSMComponentFilterManager which was causing the filter page
to appear as blank, because the page was being pinned with the wrong argument. That is also fixed.

Change-Id: Ib4bc413fcda9a5c98ae57f94e1c8a68fe9aacda3
Reviewed-on: https://asterix-gerrit.ics.uci.edu/1205
Sonar-Qube: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Taewoo Kim <wangsaeu@yahoo.com>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>