[ASTERIXDB-2288][COMP] Fix field access with UNION
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
For UNION, if there is an assign operator above UNION, the optimizer
will move the assign operator under UNION into both branches. If
the assign operator has field access by index, the index is based on
the output record of the union operator which could be different from
the field index in left and right branches. The fix is to map the
index to the correct index of the input operator where the assign
is copied to.

Change-Id: I4d65374aca6959fd5e70279fac8f7f0a6c7c64c5
Reviewed-on: https://asterix-gerrit.ics.uci.edu/3400
Contrib: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>