[ASTERIXDB-2796][RT] Fix IndexOutOfBoundsException in hash group-by
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
IndexOutOfBoundsException happens in hash group-by when WITH
clause exists. The variables from the WITH clause get propagated
through the decor variables of the group-by. The number of output
keys are the sum of both the group-by keys and decor variables, but
the hash functions & comparators used are only for the group-by keys.

- pass only the aggregated keys to the comparators and hash functions.

Change-Id: I19dc652872a1f030b6afa509711e7e0700e86856
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/8743
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>