[ASTERIXDB-2933][COMP][EXT] Pushdowns Part2: Pushdown Rule
- user model changes: no
- storage format changes: no
- interface changes: yes

Details:
Add a rule that computes the expected schema and sets it
to the DataSourceScanOperator. The computed schema is then
passed to the Parquet Reader to 'clip' the Parquet file's
schema. The resulting clipped schema is then used to tell
the reader what column should be read.

Interface changes:
- Change IProjectionInfo<List<T>> to IProjectionInfo<T>

Change-Id: If0c0d05473be72df6f08dfcbab2d25c36c71368e
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/12964
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Wael Alkowaileet <wael.y.k@gmail.com>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>