[ASTERIXDB-3143] Add SqlppAnalyzedExecutionTest
Details:
This is an extension to SqlppExecutionTest that runs
ANALYZE DATASET on every dataset that is updated via
an *.update.sqlpp file, and then continues to run
the suite. The idea is that no results should change
due to enabling CBO.

A few tests are filtered out that can't work with CBO
on, due to either bugs or known limitations.

Change-Id: Ib5e1b2efefe38332ea1f8f7328c63ace12ac9cab
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17428
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Vijay Sarathy <vijay.sarathy@couchbase.com>