[ASTERIXDB-3301][COMP] NPE for subquery with only SELECT ELEMENT
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
When a subquery consists of only SELECT ELEMENT, a NullPointerException
is encountered. The exception is due to the fact that the base/input
operator for the SUBPLAN of the subquery starts with a 'null' op that
gets replaced by a NTS op later after translating the subquery. This
works fine, but when the subquery is only a SELECT ELEMENT
expression, there are no intermediate operators translated and therefore
the root op of the subplan tries to reference the input op directly which
is 'null' at the time.

Change-Id: I5b314c014798daf4890f75998138954e3d858445
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17928
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Wail Alkowaileet <wael.y.k@gmail.com>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>