[ASTERIXDB-3391][OTH] Make DML statements cancellable
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
With this patch INSERT/UPSERT/DELETE and COPY statements are marked
cancellable. For atomic statements, the statements are cancelled as
long as the ingestion job is running i.e., ACKs from all nodes/partitions
are not received and the CC has not decided to commit the statement yet.

Change-Id: I16410ab9353c24597f77ab38ce06996fc5dfacd0
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18257
Reviewed-by: Murtadha Hubail <mhubail@apache.org>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Murtadha Hubail <mhubail@apache.org>