[ASTERIXDB-3471][COMP] Fix pushing down repeated expressions
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
When computing an expected schema node for a repeated
expression, the compiler creates a new node of type
ANY and tries to add or replace it. The issue is that
the newly created ANY node for the repeated expression
is replaceable and that fails the condition of
adding/replacing a node -- hence the thrown exception.
The add/replace node should ignore the newly created
ANY node if a node exists for the same expression.

Ext-ref: MB-62906
Change-Id: I6f25efc5e247d474ebbb6dba0b951aa0fe897ea3
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18546
Reviewed-by: Wail Alkowaileet <wael.y.k@gmail.com>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>