[ASTERIXDB-3543][COMP] Do not inline expressions containing collection record
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
Inlining expressions involving a collection's record
(e.g. field access $$col_rec.getField("x")) into the final
assign op (SELECT's assign op) can make query execution take
a long time especially if the records are not small. This is
because the $$col_rec has to be kept along in the frames from
the originating data-scan all the way through the operators
to the final assign op. This can prevent early projections
that aims to reduce the tuples size and increase the number
of tuples in a given frame.

This patch is to prevent inlining under certain conditions.
If the path from the source operator producing the expression
to the final assign op contains:
1. Operators with >= 2 inputs (e.g. JOINS).
2. ORDER-BY operator.
3. UNNEST operator.

Then we do not inline those expressions involving the data-scan
produced variables.

Ext-ref: MB-64252

Change-Id: Ia818f11e45bf66e3b4807938bfb635c70fa9c496
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/19351
Reviewed-by: Vijay Sarathy <vijay.sarathy@couchbase.com>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>