[ASTERIXDB-3611][RT] Add more stats to query profile
- user model changes: yes
- storage format changes: no
- interface changes: yes

Details:

Add frame counts and average tuples per frame to the
query profile. Also add the raw size of all tuples
forwarded from the operator.

Additionally, improve the calculation of average tuple
size to be a normal average rather than a running average.
This should be fine, because the number of bytes flowing
through and operator should be less than ~8 EB.

Furthermore, pull up all of these stats, along with the
tuple size stats, to the summarized profile in the query
plan. The averages represented in the summarized plan are
a weighted average of each operator's average, proportional
to the cardinality of that instance versus the total across
all partitions.

Ext-ref: MB-66797
Change-Id: I524cbfb1685ad3c7ed269c2db48482e5aba1ce61
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/19809
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>