[NO ISSUE][HYR][HTTP] Fix omitted access log on some netty responses
The CLFLogger had assumptions that if a DefaultFullHttpResponse was
not received, it would receive one or more DefaultHttpContent
instances followed by a LastHttpContent. In reality, Netty is free to
return a DefaultHttpContent that also implements LastHttpContent,
which causes the CLF log entry to never be emitted.

Ext-ref: MB-67407
Change-Id: I8aa6c6e8e61eb63a83dd259afbe30b08969128eb
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/20027
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Michael Blow <mblow@apache.org>
Reviewed-by: Michael Blow <mblow@apache.org>