[ASTERIXDB-3652][STO] Fixed column assembler issues
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
There were few issues related to union + objects + array.

In case of union in objects, introduced a logical barrier,
which in case of missing collects all the missing and only
propogates when it encountered all the children for the
current level, preventing extra null or missing being
propogated above.

There can be cases where a particular column is not presnet,
and the EndOfGroupAssembler uses that column as a delegate,
hence not able to correctly determine the ending of the array
leading to missing of value propogation to parent.

Ext-ref: MB-68881
Change-Id: I782d17f81d00210a4ca8673cc5fbcf556fc4758c
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/20469
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Peeyush Gupta <peeyush.gupta@couchbase.com>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>